LayoutTests/imported/w3c/web-platform-tests/uievents/click/auxclick_event-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Clicking with primary vs non-primary buttons

Double-click on the green box with a non-primary button. When using mouse any button other than the left button is non-primary. If a "PASS" result appears, the test passes; otherwise, it fails.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN auxclick event sequence received.
